Fixed a bug with dynamicYaml's ability to find keys.
The yaml document would parse fine but due to a short circuit `||` operator certain valid keys could never be retrieved off a dynamic object.

E.g.:
```yml
---
SomeProperty: 3.0
```
would always return null when accessed due to the capitalized first character.

I fixed the bug, simplified the code involved (IMHO), and added unit tests for failing (and existing passing) cases.
I preserved the pre-existing behavious that allowed the first character of the key to be case in-sensitive. I think this behaviour is a little silly but I did not want to introduce a breaking change.