Avoid definition of SerializationException on NET Standard
When a project targeting NET Standard 2.0 pulls in YamlDotNet 4.3.0 then
you get errors related to the multiple definition of
SerializationException as the NET Standard version of YamlDotNet
provides its own definition of this type, which has now been added to
NET Standard.

This is fixed by referencing System.Runtime.Serialization.Primitives
which defines the type and removing the local definition.