Switch out a new data structure
The only problem I am having is that the OnDeserialization method was called while the object was not being deserialized. It keeps occuring in two test cases. I'm in the process of working around it. More than likely has to do with OrderedDictionary using a hashtable as an underlying data structure.