Add support for private properties.
Extend the `ReadablePropertiesTypeInspector` to handle non-public
properties.

While it is possible for the user to provide custom type inspectors,
it's not easy for these to surface additional properties. An alternative
approach to the work here would be to make
`ReflectionPropertyDescriptor` protected and unseal
`ReadablePropertiesTypeInspector`. The `BuilderSkeleton` would also need to
allow the root `ITypeInspector`(s) to be customized.